pyproject.toml
[tool.poetry]
name = "tree-menu"
version = "0.1.0"
description = ""
authors = ["'Kimiyori' <'maksimkalin@mail.ru'>"]
readme = "README.md"
packages = [{include = "tree_menu"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
environs = "^9.5.0"
psycopg2-binary = "^2.9.5"
django-debug-toolbar = "^3.8.1"
django-mptt = "^0.14.0"
pylint = "^2.15.9"
flake8 = "^6.0.0"
pytest = "^7.2.0"
black = "^22.12.0"
mypy = "0.982"
django-stubs = {extras = ["compatible-mypy"], version = "^1.13.1"}
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
